requirements.txt
Python >= 3.8  
PyTorch >= 1.12  
NumPy  
scikit-learn  